fix(#240): extend SCIP staleness detection to find_referencing_symbols + get_callers
PR #236 added per-call staleness detection on `find_symbol`'s SCIP path,
but the same silent-miss class persisted on the other two SCIP-using
handlers — `find_referencing_symbols` (`tools/lsp.rs`) and
`get_callers` (`tools/graph.rs`) both kept their precise-tier 0.95+
confidence even when the on-disk index pre-dated the resolved files.
Live dogfood today (HEAD `54913a3a`, 8-day-old `index.scip`) confirmed
`get_callers` returning `confidence: 0.98, resolution: "scip"` for the
wrong caller line on `prepare_harness_session`.

Promote `detect_scip_staleness` + `ScipStaleness` from
`tools/symbols/handlers.rs` to a new shared module
`tools/scip_health.rs`, plus a centralised `scip_stale_warning_payload`
helper so the warning shape (code / message / recommended_action /
stale_files) stays identical across all three tools — callers can branch
on a single `code: scip_index_stale` signal.

Both new call sites swap to a degraded meta (confidence 0.55 +
`degraded_reason: scip_index_stale_vs_source` +
`confidence_basis: scip_precise_stale_index`) and attach the
`scip_index_stale_warning` payload when staleness is detected.

Tests:
- 5 new unit tests in `tools::scip_health::tests` (4 staleness shapes
  inherited from #236 + 1 warning-payload field check)
- 3 existing `read_signature_line_tests` continue to pass
- `find_symbol` integration unaffected — same shared helpers
- Full bin nextest 729 pass; clippy clean under both
  `--features http,semantic` and `--no-default-features`

Closes #240.
