+++ crates/codelens-mcp/src/tools/graph.rs
@@ -411,8 +411,32 @@ pub fn get_callers_tool(state: &AppState, arguments: &serde_json::Value) -> Tool
             }
         }
 
-        let (resolution_summary, confidence_basis, meta) =
+        let (resolution_summary, computed_basis, computed_meta) =
             call_graph_analysis(value.iter().map(|entry| entry.resolution));
+        // Issue #240: when a SCIP-resolved caller line points at a file
+        // that's been edited since the index was built, the precise-tier
+        // confidence (≥ 0.95) lies the same way `find_symbol` did
+        // pre-#236. Probe staleness against every caller file (cheap —
+        // mtime per file) and degrade meta + emit warning when any is
+        // newer than the index.
+        #[cfg(feature = "scip-backend")]
+        let scip_staleness = {
+            let candidate_files: Vec<String> = value.iter().map(|c| c.file.clone()).collect();
+            crate::tools::scip_health::detect_scip_staleness(
+                state.project().as_path(),
+                &candidate_files,
+            )
+        };
+        #[cfg(not(feature = "scip-backend"))]
+        let scip_staleness: Option<()> = None;
+        let (confidence_basis, meta) = if scip_staleness.is_some() {
+            (
+                "scip_precise_stale_index",
+                crate::tool_evidence::meta_degraded("scip", 0.55, "scip_index_stale_vs_source"),
+            )
+        } else {
+            (computed_basis, computed_meta)
+        };
         let evidence = crate::tool_evidence::tool_evidence(
             "call_graph",
             &meta,
@@ -429,6 +453,15 @@ pub fn get_callers_tool(state: &AppState, arguments: &serde_json::Value) -> Tool
             "resolution_summary": resolution_summary,
             "evidence": evidence,
         });
+        #[cfg(feature = "scip-backend")]
+        if let Some(stale) = scip_staleness.as_ref()
+            && let Some(map) = payload.as_object_mut()
+        {
+            map.insert(
+                "scip_index_stale_warning".to_owned(),
+                crate::tools::scip_health::scip_stale_warning_payload(stale),
+            );
+        }
         if !unknown_args.is_empty()
             && let Some(map) = payload.as_object_mut()
         {
